pyproject.toml
[tool.poetry]
name = "homework-27"
version = "0.1.0"
description = ""
authors = ["Andrey Kopylov <104315220+obscurezz@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "homework_27"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
django-import-export = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
